requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
python-dotenv==1.0.1
openai>=1.60.0,<2.0.0
httpx==0.27.2
markdown==3.7

pydantic